pyproject.toml
[tool.poetry]
name = "home-monitoring-display"
version = "0.1.0"
description = "Display of home_monitoring metrics"
authors = ["tdurouchoux <thibault.durouchoux@gmail.com>"]
readme = "README.md"
packages = [{include = "home_monitoring_display"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
influxdb-client = "^1.36.1"
PyYAML = "^6.0"
panel = "0.14.4"
streamz = "^0.6.4"
reactivex = "^4.0.4"
pandas = "^1.5.3"
pyparsing = "^3.0.9"
scipy = {url = "https://www.piwheels.org/simple/scipy/SciPy-1.8.1-cp39-cp39-linux_armv6l.whl#sha256=f6865c8d400b37c65af13b1d7751480840f99169da80478095508f2929365e39"}
influxdb = "^5.3.1"
pytz = "^2023.3"
scikit-learn = "^1.2.2"
contourpy="1.0.7"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.group.panel]
optional = true

[tool.poetry.group.panel.dependencies]
hvplot = "^0.8.3"
holoviews = "^1.15.4"
panel = "^0.14.4"
matplotlib = "^3.7.1"

[tool.poetry.group.inky]
optional = false

[tool.poetry.group.inky.dependencies]
inky = "^1.5.0"
Pillow = "^10.0.0"
buttonshim="^0.0.2"
smbus = "^1.1.post2"


[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple"
priority = 'default'

[[tool.poetry.source]]
name = "PyPI"
priority = 'primary'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
